requirement.txt
 streamlit==1.35.0
streamlit-lottie==0.0.5
PyPDF2==3.0.1
google-generativeai==0.4.1
python-dotenv==1.0.1
requests==2.31.0
